Add OSP command get_vts.
This adds the command get_vts, for getting the list and details about
vulnerability tests a OSP scanner might offer.

So far, the tests have only an id and a name. It is intended
to extend this by some standard elements which are shared among
different scanners and with individual elements which are very specific
for a scanner.

With this status, it would be possible to have a client get the list and
allow the user to select one or many tests that are desired.

Follow-ups will be to add receive the vt selection from the client in
order to execute only these.

This change uses a simple dict for storage. With extending
capabilities, it might be helpful to use advanced containers/collections.

* ospd/ospd.py: Add command get_vts and handling of data.

* doc/OSP.xml: Add documentation for command get_vts.