Add element "vts" to parameters for starting scans
The element "vts" is added to the base scanner parameters
and allows to send a list of vt's to be executed by the scanner.

The scanner needs to parse and handle the list on its own.

Since "vts" is a comma-separated list of vt_ids, to prevent
parsing/escaping problems, the syntax of a vt_id is now
limited to certain characters which are typically used for ids.