Add status type, sort data types
This commit adds a definition for the `status` data type which was used
but never defined.

It also makes it clear that all `status` attributes follow this type,
which was inconsistent before.

The definition of the `boolean` data type is updated to have the
pattern match the description.

Lastly, the data types are sorted alphabetically to improve readability.