Improve connection handling
This commit improves the way connections are handled by the
`handle_client_stream()` method.

Previously, data from the client socket was read until a timeout was
reached after two seconds, i.e. when there was no activity from the open
client socket for two seconds. Beyond reaching this timeout or
encountering an error, there was no way to exit the `while True:` loop
containing the `recv` and `read` calls.

This lead to problems when a client sent data and waited less than two
seconds before closing the connection. Since the timeout was never
reached, the `data.append()` method continued to append `bytes` to the
`data` array indefinitely.

This commit introduces a conditional to check if any additional data was
read and to `break` the loop if not, i.e. all data sent by the client
was consumed.

This has the added benefit of reducing the time the client has to wait
after it is done sending by the two seconds the server previously waited
before starting to process the data.