Simplify buffer check
This commit further simplifies the check for received data by inverting
the conditional and taking into account that an empty array is `False`.

It also removes the useless conditional checking the length of the
`data` array. The `data` array cannot be zero-length at this point
because at least an empty `bytes` object will have been appended to it,
resulting in a non-zero length.