Set minimum version for python-redis in setup.py.
The 'count' argument that is used for accessing redis was introduced
in python-redis >= 3.0.1. This patch adds this minimum requirement.
It made my setup work again on a Debian 9 when running setup.py directly
for installation.

Also it is a documentation of the minimum requirement, I found it
in no other place.