Implement a new command infrastructure
All ospd commands are now representated by an instance of a BaseCommand
class. This BaseCommand class automatically registers a command class.

Currently the command classes handle a xml request and are calling the
OSPDDaemon for further information or processing. In future the
"interface" between the command classes and the daemon should be
stripped down.