Delay creation of multiprocessing manager
When running as daemon the manager must be created within the background
process. If the manager is started early it is running in the non forked
process which will exit after the forking. Therefore the manager will be
shut down too. To avoid this the creating of the manager must be delayed
to be started withing the forked background process.