Fix get_scanner_details(). Running the test with python3 -m setup.py tests, a test was failing while running with python3 -m unittest, it doesn't.
The scanner_params variable was overwritten with the xml Element and
passed dictionary was lost.

A second issue was the et.Element type. isinstance() in xml.simple_response_str()
does not recognize et.Element as Element, and therefore the parameter was not added to the final
response. For this reason, Element is used in protocol instead of et.Element.