handle dead hosts
When the test_alive_hosts_only feature of openvas is enabled we
do not scan every single host in the target list and therfore
do not have the host progress for all hosts. To still be able
to update the progress bar we need to send all the dead hosts
from openvas to ospdopenvas. This is not an optimal solution as
we get big jumps in the progress bar if the target lists has many
hosts which are dead because they all get sent at once after the
host discovery is completed.