Add/modify methods to handle finishes hosts.
A host can be added as dead directly to the dead host count in the scan table
or can be set as dead because the scan progress of the host is -1

A scanner, like openvas, can scan a host and set the host as dead at the end.
In this case in which a host is alive until it is recognize as dead, this options
are useful.