Handle denied hosts.
Denied host are not scanned by openvas, but a child process to check if
it is allowed/denied is started. No ip address for this hosts is
written in kb and therefore most be handled in a special way,
The error message received from openvas contains the ip address
and it taken from here.