Use available memory instead of free memory
When check for enough free memory before starting a scan, check for available memory now

When the command line 'free -h' is called, the output shows two different values: available and free memory. Available is the one with more fluctuation and the one we should check instead of free, because it gives a more realistic value for checking if there is enough memory before starting a queued scan.

More info can be found in the free command man page.