Fix get_count_total(). Accept -1 value set by the server.
The value set by the server has priority over the value
calculated from the original target list by ospd.
As ospd is not intelligent enough to check the amount of valid
hosts, check for duplicated or invalid hosts, consider a negative
value set for the server, in case it detects an invalid target string
or a different amount than the orignal amount in the target list.