Wait a minute between scans starts.
Currently the queue system in ospd prevents to start a new scan if there is not enough available memory. This produces that new scans are put in the queue, while the memory condition is not satisfied.
Once some memory is released, all scans in queue are started. In the issue mentioned above, I pasted what I found in the log.

This avoids ospd to start many queued scans simultaneously. So the last started scan has time enough to take some memory (launch some child processes).