Fix timeout preference handling
Scripts timeouts are special preferences and in the openvas side are
treated in a especial way too.

The timeout preference is in the 13th position in the script metadata in redis,
and not with other preferences.
Also, the timeout is stored in globals with format timeout.<SCRIPT-OID>|||<value>,
while other script preferences have the form oid:name:id:type|||value.

The user is allowed to send a custom timeout, even when the script has no default one
set in the description block (use default 320sec). When the user send a custom timeout
the preference is not found and ospd-openvas logs an error message.

With this patch, the preference with ID:0 (timeout) is always handled and sent to openvas.