Fix resume task.
In some cases, the task is stopped and quickly resumed. In this cases there was race condition
in which the scan id is present in the scan table but there is no target information.
Although the quick resume can lead first into a stopped and scan followed by an interrupted scan,
finally the scan can be resumed.