Add: notus advisories in get_vts
When calling get_vts ospd-openvas will try to load the notus advisories;
look them up and yielding the response; when it is not in the notus
advisories it will return the information based on redis from openvas as
previously.

With this integration a client e.g. gvmd should get all needed
information for notus results.