Refactor: use redis cache, fix notus integration
The notus advisories are huge lists containing thousands of vts
therefore it is better to store them in a redis cache than loading and
iterating through those for each get_nvts. Therefore this commits adds
redis for notus.

Before the actual data was directly used which in return was unparseable
for the XML encoder there it is also chaning vthelper to deal with that.