Add: notus advisories in get_vts (#518)
Add: notus advisories in get_vts

When calling get_vts ospd-openvas will try to load the notus advisories;
look them up and yielding the response; when it is not in the notus
advisories it will return the information based on redis from openvas as
previously.

With this integration a client e.g. gvmd should get all needed
information for notus results.

The notus advisories are huge lists containing thousands of vts
therefore it is better to store them in a redis cache than loading and
iterating through those for each get_nvts. 

This results in a longer ramp up time favoring faster resolve time.

Co-authored-by: Björn Ricks <bjoern.ricks@gmail.com>