Fix: stacktrace on sensor when unexpectedly closing a ssh connection
In the circumstances that ospd-openvas is run on a sensor and while
starting a scan a ssh service is disabled it should not print a
BrokenPipeError based stacktrace triggered by subprocess of start_scan.

Since this is self correcting occurance it prints a warning instead.

(cherry picked from commit 487a58e6e100d275a562eeedcc7f85c620362378)