Fix: Stop and resume scan (#604)
* Fix dead lock

When the scan is started and immediately stopped it keeps looping and waiting for the already finished openvas process.
When a scan is started, stopped when it is still 0% progress and quickly resumed, the task ended as interrupted, or not in a blocking situation with "requested" status.

* Handle the case in which the scan was already deleted.

* Speed up the stop scan when the task has just started.
This allows to stop the plugins loadup earlier.

For testing:
1a) Start a vulnerability scan against a single host target
2a) As soon as the scan task has the status "0%", stop the scan task
3a) Once the scan task has the status "Stopped at 0%", resume the scan task
4a) If the scan task has the status "Stopped at 0%" or "Interrupted at 0%" subsequently, resume the scan task again

With the patch, it is possible to do this and the scan is stopped successfully and it is possible to resume it with out issues.

(cherry picked from commit 41b25c2974a9afddd12cb130e87854fb40ea5f56)

# Conflicts:
#	ospd/command/command.py
#	ospd_openvas/preferencehandler.py