Fix: missing reload on update_vts
On check_feed update_vts is triggered however it just updates openvas
vts and not notus; this commit changes that by reloading the notus cache
as well.

Although ospd-openvas should handle the calls reload-cache of notus is
now checking if there is already a reload process called and blocking
until that one is finished to prevent multiple reloads at the same time.