Refactor: refactoring smoketests
To be able to quickly adapt and create tests for ospd-openvas the
smoketest go project is refactored to be capable of running those tests
and print if it is successful or not and after the complete run print
details of failures if any.

In addition those test will be run within the new workflow
smoketests.yml.