Refactor: add possibility to test scan-configs
To be able to test scan-configs smoketest are extended to parse nasl
files for oid, family and includes.

In extension it is also parsing scan-configs.

For feed preparation there is a new cmd `feed-preparer` which parses
scan-configs and nasl files to lookup the needed files for the
given policies and copies them into a target folder.

The actual test case in `policy/init.go` does start the policy
`Discovery` and verify if authenticated is enabled afterwards.