Fix: race condition within daemon and notus
While creating a hash sum verification it can happen that a lock exist
within daemon which is not yet resolved while already doing a hashsum
verification in a different process.

This results in a non functioning OSPD since it stuck on loading; to
prevent a scenario like this notus will create the verification only
when needed and not eagerly while creating the first instance.