Fix: use `~/.gnupg` when `/etc/openvas/gnupg` is not available
Instead of just using GOS defaults `/etc/openvas/gnupg` check if the
directory exists and when not use `$HOME/.gnupg` instead.

If both are not available print a warning that the env variable
GNUPGHOME should be set but stick with the failing `$HOME/.gnupg` to
prevent None checking.

Fixes https://github.com/greenbone/ospd-openvas/issues/765