Uses a new and empty Redis namespace for the Notus cache. (#772)
Uses a new and empty Redis namespace for the Notus cache.

What:
Uses a new and empty Redis namespace for the Notus cache.

Why:
It was using the redis namespace 0, which is reserved for openvas to keep a list of in-use kbs.

How:
$ redis-cli -s

    info namespaces
    keys *

(cherry picked from commit 72d4f4653b5b72ecf3fe8b8612b5e80f72387f1a)