Separate notus redis cache from openvas cache (#773)
* Update db.get_filenames_and_oids()
Now receives a pattern and a parser for searching keys.
This allows that the method can be use in different cache, with different structures
to find keys with different patterns.

* Add get_oids() to Notus class

* Nvti.get_oids() only returns the openvas vts.

* Add comments in the code. Basically, preference handler doesn't send oid that belongs to Notus. Therefore, only get nvti oids

* Initialized the OpenvasVtsFilter class with the notus cache too.
The notus Cache class has it own method to get oids and we use it.

* It cache class (nvti and notus) have their own methods to get oids.
We use it separatelly to create the vts collection.

* Remove force_reload() method, it is never used/called. Also, remove related tests

* Dont get the family of notus advisories.

Family is only use to process vts groups to be sent to openvas. As notus advisories are handled by notus and not by openvas, it makes no sense to get the advisories and then removed from the plugin list.

* Finally, we can make NVTICache independent of Notus.

* Make Notus optional in VtHelper

It could be that we are only interested in a VtHelper for nvts and not for notus advisories.

* Make PreferenceHandler independent of notus

* Fix/update tests.

Co-authored-by: ArnoStiefvater <ArnoStiefvater@users.noreply.github.com>
(cherry picked from commit 5b25f14f4fd05338ea9ef503b398a926caa64974)