Fix: queued scan count (#850)
Wait 0, 10, 20, 30, ..., or 90 miliseconds when add a new scan in the queue.
This try to avoid a race condition counting the amount of queued scans when many task
are started at the same time. This leads into wrong queue position
number shown in the logs and could jump the max_queued_scans setting
in some cases.